+++ packages/oauth/oauth-provider/src/router/assets/assets.ts
@@ -1,7 +1,8 @@
 import type { HydrationData as FeHydrationData } from '@atproto/oauth-provider-frontend/hydration-data'
 import type { HydrationData as UiHydrationData } from '@atproto/oauth-provider-ui/hydration-data'
-import { combineMiddlewares } from '../lib/http/middleware.js'
-import { Simplify } from '../lib/util/type.js'
+import { CspConfig } from '../../lib/csp/index.js'
+import { combineMiddlewares } from '../../lib/http/middleware.js'
+import { Simplify } from '../../lib/util/type.js'
 import { parseAssetsManifest } from './assets-manifest.js'
 
 // If the "ui" and "frontend" packages are ever unified, this can be replaced
@@ -32,3 +33,22 @@ export const assetsMiddleware = combineMiddlewares([
   ui.assetsMiddleware,
   fe.assetsMiddleware,
 ])
+
+export const SPA_CSP: CspConfig = {
+  // API calls are made to the same origin
+  'connect-src': ["'self'"],
+  // Allow loading of PDS logo & User avatars
+  'img-src': ['data:', 'https:'],
+  // Prevent embedding in iframes
+  'frame-ancestors': ["'none'"],
+}
+
+/**
+ * @see {@link https://docs.hcaptcha.com/#content-security-policy-settings}
+ */
+export const HCAPTCHA_CSP: CspConfig = {
+  'script-src': ['https://hcaptcha.com', 'https://*.hcaptcha.com'],
+  'frame-src': ['https://hcaptcha.com', 'https://*.hcaptcha.com'],
+  'style-src': ['https://hcaptcha.com', 'https://*.hcaptcha.com'],
+  'connect-src': ['https://hcaptcha.com', 'https://*.hcaptcha.com'],
+}
