+++ packages/oauth/oauth-provider/src/router/create-account-page-middleware.ts
@@ -1,32 +1,34 @@
 import type { IncomingMessage, ServerResponse } from 'node:http'
-import { ActiveDeviceSession } from '@atproto/oauth-provider-api'
+import type { ActiveDeviceSession } from '@atproto/oauth-provider-api'
 import {
+  Middleware,
   Router,
   validateFetchDest,
   validateFetchMode,
   validateOrigin,
 } from '../lib/http/index.js'
 import type { OAuthProvider } from '../oauth-provider.js'
-import type { RouterOptions } from './router-options.js'
-import { sendAccountPageFactory } from './ui-router/send-account-page.js'
+import { sendAccountPageFactory } from './assets/send-account-page.js'
+import { sendErrorPageFactory } from './assets/send-error-page.js'
+import type { MiddlewareOptions } from './middleware-options.js'
 
-export function accountRouter<
+export function createAccountPageMiddleware<
   T extends object | void = void,
   TReq extends IncomingMessage = IncomingMessage,
   TRes extends ServerResponse = ServerResponse,
 >(
   server: OAuthProvider,
-  options: RouterOptions<TReq, TRes>,
-): Router<T, TReq, TRes> {
-  const { onError } = options
+  { onError }: MiddlewareOptions<TReq, TRes>,
+): Middleware<T, TReq, TRes> {
   const sendAccountPage = sendAccountPageFactory(server.customization)
+  const sendErrorPage = sendErrorPageFactory(server.customization)
 
   const issuerUrl = new URL(server.issuer)
   const issuerOrigin = issuerUrl.origin
 
   const router = new Router<T, TReq, TRes>(issuerUrl)
 
-  router.get<never>(/^\/account(?:\/.*)?$/, async function (req, res, next) {
+  router.get<never>(/^\/account(?:\/.*)?$/, async function (req, res) {
     try {
       res.setHeader('Referrer-Policy', 'same-origin')
 
@@ -41,7 +43,7 @@ export function accountRouter<
       const deviceAccounts =
         await server.accountManager.listDeviceAccounts(deviceId)
 
-      return sendAccountPage(req, res, {
+      sendAccountPage(req, res, {
         deviceSessions: deviceAccounts.map(
           (deviceAccount): ActiveDeviceSession => ({
             account: deviceAccount.account,
@@ -57,9 +59,11 @@ export function accountRouter<
         `Failed to handle navigation request to "${req.url}"`,
       )
 
-      next(err)
+      if (!res.headersSent) {
+        sendErrorPage(req, res, err)
+      }
     }
   })
 
-  return router
+  return router.buildMiddleware()
 }
