+++ packages/oauth/oauth-provider/src/router/middleware-options.ts
@@ -1,7 +1,7 @@
 import type { IncomingMessage, ServerResponse } from 'node:http'
 import type { ErrorHandler } from './error-handler.js'
 
-export type RouterOptions<
+export type MiddlewareOptions<
   Req extends IncomingMessage = IncomingMessage,
   Res extends ServerResponse = ServerResponse,
 > = {
